+++ Automate/Framework/AutomationFactory.cs
@@ -38,6 +38,12 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <summary>The tile area on the farm matching the shipping bin.</summary>
         private readonly Rectangle ShippingBinArea = new Rectangle(71, 14, 2, 1);
 
+        /// <summary>Whether the Better Junimos mod is installed.</summary>
+        private readonly bool HasBetterJunimos;
+
+        /// <summary>Whether the Deluxe Auto-Grabber mod is installed.</summary>
+        private readonly bool HasDeluxeAutoGrabber;
+
 
         /*********
         ** Public methods
@@ -47,7 +53,9 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <param name="automateShippingBin">Whether to treat the shipping bin as a machine that can be automated.</param>
         /// <param name="monitor">Encapsulates monitoring and logging.</param>
         /// <param name="reflection">Simplifies access to private game code.</param>
-        public AutomationFactory(ModConfigObject[] connectors, bool automateShippingBin, IMonitor monitor, IReflectionHelper reflection)
+        /// <param name="hasBetterJunimos">Whether the Better Junimos mod is installed.</param>
+        /// <param name="hasDeluxeAutoGrabber">Whether the Deluxe Auto-Grabber mod is installed.</param>
+        public AutomationFactory(ModConfigObject[] connectors, bool automateShippingBin, IMonitor monitor, IReflectionHelper reflection, bool hasBetterJunimos, bool hasDeluxeAutoGrabber)
         {
             this.Connectors = connectors
                 .GroupBy(connector => connector.Type)
@@ -55,6 +63,8 @@ namespace Pathoschild.Stardew.Automate.Framework
             this.AutomateShippingBin = automateShippingBin;
             this.Monitor = monitor;
             this.Reflection = reflection;
+            this.HasBetterJunimos = hasBetterJunimos;
+            this.HasDeluxeAutoGrabber = hasDeluxeAutoGrabber;
         }
 
         /// <summary>Get a machine, container, or connector instance for a given object.</summary>
@@ -70,7 +80,7 @@ namespace Pathoschild.Stardew.Automate.Framework
 
             // machine
             if (obj.ParentSheetIndex == 165)
-                return new AutoGrabberMachine(obj, location, tile);
+                return new AutoGrabberMachine(obj, location, tile, ignoreSeedOutput: this.HasDeluxeAutoGrabber);
             if (obj.name == "Bee House")
                 return new BeeHouseMachine(obj, location, tile);
             if (obj is Cask cask)
@@ -159,7 +169,7 @@ namespace Pathoschild.Stardew.Automate.Framework
         {
             // machine
             if (building is JunimoHut hut)
-                return new JunimoHutMachine(hut, location);
+                return new JunimoHutMachine(hut, location, ignoreSeedOutput: this.HasBetterJunimos);
             if (building is Mill mill)
                 return new MillMachine(mill, location);
             if (this.AutomateShippingBin && building is ShippingBin bin)