+++ Automate/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added support for blocking input or output for a chest (thanks to kice!).
+* Fixed compatibility with Auto-Grabber Mod and Better Junimos. If they're installed, seeds in auto-grabbers (Auto-Grabber Mod) or Junimo huts (Better Junimos) will be ignored.
 * Fixed "don't use this chest for automation" option not taking effect immediately.
 
 ## 1.11.1